Addressed copilot review comments
